# internal/api/docs/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      error:
        example: descriptive error message
        type: string
    type: object
...
  models.Backlink:
    properties:
      bear_id:
        type: string
      bear_raw:
        type: string
      id:
        type: string
      linked_by_id:
        description: note that contains the link
        type: string
      linking_to_id:
        description: note being linked to
        type: string
      location:
        type: integer
      modified_at:
        type: string
      title:
        type: string
      version:
        type: integer
    type: object
...
  models.WriteQueueItem:
    properties:
      action:
        description: create | update | add_tag | trash | add_file | archive | rename_tag
          | delete_tag
        type: string
      applied_at:
        type: string
      consumer_id:
        type: string
      created_at:
        type: string
      error:
        type: string
      id:
        type: integer
      idempotency_key:
        type: string
      lease_until:
        type: string
      note_id:
        type: string
      note_sync_status:
        description: |-
          NoteSyncStatus is populated during lease to indicate the associated note's sync_status.
          Not stored in DB — filled by JOIN during LeaseQueueItems.
        type: string
      payload:
        description: JSON with operation data
        type: string
      processing_by:
        type: string
      status:
        description: pending | processing | applied | failed
        type: string
    type: object
info:
  contact: {}
  description: API for reading and writing Bear notes via the hub.
  title: Bear Sync Hub API
  version: "1.0"
paths:
  /api/attachments/{id}:
    get:
      description: Downloads the attachment file by ID. Returns the file with Content-Disposition
        header.
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an attachment
      tags:
      - Attachments
...
  /api/notes/{noteID}/attachments:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a file and attaches it to a note. Max file size is 5 MB.
        Requires Idempotency-Key header.
      parameters:
      - description: Note ID
        in: path
        name: noteID
        required: true
        type: string
      - description: Idempotency key for deduplication
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: File to attach
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.WriteQueueItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Attach a file to a note
      tags:
      - Notes
  /api/notes/{noteID}/backlinks:
    get:
      description: Returns all notes that link to the specified note.
      parameters:
      - description: Note ID
        in: path
        name: noteID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Backlink'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List backlinks for a note
      tags:
      - Backlinks
...
securityDefinitions:
  BearerAuth:
    description: 'Consumer Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"